Put in governance page and disable menu when no stakepool is found
