+++ app/components/AgendaOverview/index.js
@@ -9,6 +9,7 @@ class AgendaOverview extends React.Component {
     super(props);
     this.state = {
       selectedChoiceId: this.props.selectedChoice,
+      disabled: this.props.disabled,
     };
   }
 
@@ -20,7 +21,7 @@ class AgendaOverview extends React.Component {
 
   render() {
     const { agenda, closeCurrentAgenda } = this.props;
-    const { selectedChoiceId } = this.state;
+    const { selectedChoiceId, disabled } = this.state;
     const { setSelecedChoiceId, updatePreferences } = this;
     const activeChoiceId = this.props.selectedChoice;
     const agendaId = agenda.getId();
@@ -41,7 +42,8 @@ class AgendaOverview extends React.Component {
           choices,
           setSelecedChoiceId,
           updatePreferences,
-          closeCurrentAgenda
+          closeCurrentAgenda,
+          disabled
         }}
       />
     );
